requirements.txt
tensorflow
scipy
mtcnn
keras_facenet
opencv-python
numpy
Pillow
streamlit
streamlit_webrtc
av
